+++ Sources/SwiftUIKit/Presentation/SheetContext.swift
@@ -28,8 +28,9 @@ import SwiftUI
  }
  ```
  
- You can also setup a global context in the application root,
- and pass it as an environment object, to get a single value.
+ This view modifier will also inject the provided context as
+ an environment object into the view hierarchy, to let other
+ views in the same view hierarchy reuse the same context.
  */
 public class SheetContext: PresentationContext<AnyView> {
     
@@ -41,10 +42,13 @@ public class SheetContext: PresentationContext<AnyView> {
 public extension View {
     
     /// Bind an ``SheetContext`` to the view.
+    ///
+    /// This also injects this context as environment object.
     func sheet(_ context: SheetContext) -> some View {
         sheet(
             isPresented: context.isActiveBinding,
             content: context.content ?? EmptyView().any
         )
+        .environmentObject(context)
     }
 }